refactor : 도메인 내에서 view 로직 분리
