--- src/main/java/game/util/ResultView.java
@@ -1,8 +1,0 @@
-package game.util;
-
-public class ResultView {
-
-    public static void printPlainMessage(final String message) {
-        System.out.println(message);
-    }
-}
