+++ src/main/java/game/view/InputView.java
@@ -1,11 +1,11 @@
-package game.util;
+package game.view;
 
 import java.util.Scanner;
 
 public class InputView {
     static Scanner scanner = new Scanner(System.in);
 
-    public static String insertValue() {
-        return scanner.nextLine();
+    public static int insertInt() {
+        return scanner.nextInt();
     }
 }
